Drop control name BC, added framework BC
